requirements.txt
aenum==3.1.16
certifi==2025.11.12
cfbd==5.13.2
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pydantic==1.10.24
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
xgboost==3.1.2
